backend/requirements.txt
Django==4.2.23
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
openai==1.90.0
python-decouple==3.8
whitenoise==6.6.0
django-cors-headers==4.7.0
django-extensions==4.1.0
mongoengine==0.29.1
pymongo==3.11.4
Pillow==11.2.1
